Fix broken plugin structure
