+++ plugins/azure-blob-storage/CMakeLists.txt
@@ -17,4 +17,7 @@ find_package(Tenzir REQUIRED)
 
 TenzirRegisterPlugin(
   TARGET azure-blob-storage
-  ENTRYPOINT src/plugin.cpp)
+  ENTRYPOINT src/plugin.cpp
+  INCLUDE_DIRECTORIES include
+  BUILTINS GLOB "${CMAKE_CURRENT_SOURCE_DIR}/builtins/*.cpp")
+
